// src/app/shared/services/api/service-proxies.ts
export class ResponseBase implements IResponseBase {
    status!: string;
    message?: string;

    constructor(data?: IResponseBase) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.status = data["status"];
            this.message = data["message"];
        }
    }

    static fromJS(data: any): ResponseBase {
        data = typeof data === 'object' ? data : {};
        let result = new ResponseBase();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["status"] = this.status;
        data["message"] = this.message;
        return data;
    }
}

export interface IResponseBase {
    status: string;
    message?: string;
}
// ...
export class LoginRequest implements ILoginRequest {
    authType?: string;
    authCredentials?: AuthCredentials;

    constructor(data?: ILoginRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
            this.authCredentials = data.authCredentials && !(<any>data.authCredentials).toJSON ? new AuthCredentials(data.authCredentials) : <AuthCredentials>this.authCredentials;
        }
    }

    init(data?: any) {
        if (data) {
            this.authType = data["authType"];
            this.authCredentials = data["authCredentials"] ? AuthCredentials.fromJS(data["authCredentials"]) : <any>undefined;
        }
    }

    static fromJS(data: any): LoginRequest {
        data = typeof data === 'object' ? data : {};
        let result = new LoginRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["authType"] = this.authType;
        data["authCredentials"] = this.authCredentials ? this.authCredentials.toJSON() : <any>undefined;
        return data;
    }
}

export interface ILoginRequest {
    authType?: string;
    authCredentials?: IAuthCredentials;
}
// ...
export class LoginResponse extends ResponseBase implements IAnonymous22 {
    responseData?: ResponseData4;

    constructor(data?: IAnonymous22) {
        super(data);
        if (data) {
            this.responseData = data.responseData && !(<any>data.responseData).toJSON ? new ResponseData4(data.responseData) : <ResponseData4>this.responseData;
        }
    }

    init(data?: any) {
        super.init(data);
        if (data) {
            this.responseData = data["responseData"];
        }
    }

    static fromJS(data: any): LoginResponse {
        data = typeof data === 'object' ? data : {};
        let result = new LoginResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["responseData"] = this.responseData;
        super.toJSON(data);
        return data;
    }
}

export interface IAnonymous22 extends IResponseBase {
    responseData?: IResponseData4;
}
// ...
export class AuthCredentials implements IAuthCredentials {
    username?: string;
    password?: string;

    constructor(data?: IAuthCredentials) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.username = data["username"];
            this.password = data["password"];
        }
    }

    static fromJS(data: any): AuthCredentials {
        data = typeof data === 'object' ? data : {};
        let result = new AuthCredentials();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["username"] = this.username;
        data["password"] = this.password;
        return data;
    }
}

export interface IAuthCredentials {
    username?: string;
    password?: string;
}
// ...
export class ResponseData4 implements IResponseData4 {

    constructor(data?: IResponseData4) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
    }

    static fromJS(data: any): ResponseData4 {
        data = typeof data === 'object' ? data : {};
        let result = new ResponseData4();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        return data;
    }
}

export interface IResponseData4 {
}